animation smooth and hero component added fade effect added
