+++ src/App.css
@@ -210,7 +210,6 @@ justify-content:flex-end;
   mix-blend-mode: exclusion;
   background-color: #fff;
 }
-
 #nav {
   display: flex;
   align-items: center;
@@ -250,9 +249,19 @@ justify-content:flex-end;
   width: 100%;
   height: 100vh;
 }
+.arr{
+  transition: all ease-in-out 0.4s;
+}
+.op:hover>.arr{
+  transform: rotate(45deg);
+  transition: all ease-in-out 0.4s;
+}
+#o{
+  border-radius: 50%;
+}
 @media only screen and (max-width: 600px) {
  #hero {
-  height: 80vh;
+  height: 83vh;
   }
   
   #main{
@@ -271,7 +280,7 @@ justify-content:flex-end;
   right: 0px;
 }
 .wrapper{
-  top: 38vh;
+  top: 41vh;
 }
 .moi{
  font-size:20px !important;
@@ -471,10 +480,13 @@ padding-top: 25vh !important;
   opacity: .7;
  /* -webkit-text-stroke: 4px #fff; */
 }
+.elem>h1{
+  transition: all 0.5s ease-in-out;
+}
 .elem:hover>h1{
-  margin-left:10px;
-  opacity: 0.2;
-  transition: all 1s ease-in-out;
+  margin-left:17px;
+  opacity: 0.3;
+  transition: all 0.7s ease-in-out;
 }
 #second {
   color: #fff;
