+++ src/App.jsx
@@ -1,19 +1,15 @@
 import './App.css';
-import React, { useEffect,useState} from "react";
+import React, { useEffect,useState,useRef} from "react";
+import Hero from './components/Hero';
 import gsap from "gsap";
 import { ScrollTrigger } from 'gsap/ScrollTrigger';
 import {IoArrowDownSharp,IoChatbubbleEllipses} from "react-icons/io5";
-import { FiArrowUpRight } from "react-icons/fi";
-import loh from './loh.jpeg';
 import LocomotiveScroll from 'locomotive-scroll';
 import Navbar,{scrollAnimation} from './components/Navbar';
-import HeroBanner from './components/HeroBanner';
-import Clients from './components/Clients';
-import About from './components/About';
 import Footer from './components/Footer';
-import Preloader from './components/Preloader';
 import SideBar from './components/SideBar';
 import Transition from './transition/Transition';
+
 gsap.registerPlugin(ScrollTrigger);
 
 function App() {
@@ -101,98 +97,7 @@ function App() {
       window.removeEventListener('mousemove', circleChaptaKaro);
       clearTimeout(timeout);
     };
-  }, [mobileMenu]);
-  useEffect(() => {
-    // if(loading)return
-    function firstPageAnim() {
-      let tl = gsap.timeline();
-
-      tl.from("#nav", {
-        y: -10,
-        opacity: 0,
-        duration: 1.5,
-        delay: -1,
-        ease: "expo.inOut",
-      })
-        .to(".boundingelem", {
-          y: 0,
-          ease: "expo.inOut",
-          duration: 2,
-          delay: -1,
-          stagger: 0.2,
-        })
-        .from("#herofooter", {
-          y: -10,
-          opacity: 0,
-          duration: 1.5,
-          delay: -1,
-          ease: "expo.inOut",
-        });
-    }
-
-    firstPageAnim();
-  }, []);
-  useEffect(()=>{
-    function enterAnimation(link, e, index) {
-      link.tl.tweenFromTo(0, "midway");
-    }
-    
-    // Mouseleave function
-    function leaveAnimation(link, e) {
-      link.tl.play();
-    }
-    
-   
-    // Get all links
-    let workLinks = document.querySelectorAll(".js-work-link");
-    
-    workLinks.forEach((link, index, value) => {
-      let underline = link.querySelector(".underline");
-      link.tl = gsap.timeline({paused: true});
-      
-      link.tl.fromTo(underline, {
-        width: "0%",
-        left: "0%",
-      }, {
-        width: "100%",
-        duration: 0.2,
-      });
-      
-      link.tl.add("midway");
-      
-      link.tl.fromTo(underline, {
-        width: "100%",
-        left: "0%",
-      }, {
-        width: "0%",
-        left:"100%",
-        duration: 0.2, 
-        immediateRender: false
-      });
-    
-      // Mouseenter
-      console.log(link.tl);
-      link.addEventListener("mouseenter", (e) => {
-        enterAnimation(link, e, index);
-      });
-    
-      // Mouseleave
-      link.addEventListener("mouseleave", (e) => {
-        leaveAnimation(link, e);
-      });
-      return ()=>{
-        link.removeEventListener("mouseenter", (e) => {
-          enterAnimation(link, e, index);
-        });
-      
-        // Mouseleave
-        link.removeEventListener("mouseleave", (e) => {
-          leaveAnimation(link, e);
-        });
-      }
-    });
-  },[])
-  
+  }, [mobileMenu]);  
   useEffect(()=>{
     if(window.innerWidth<1400){
       setMobileMenu(true);
@@ -215,44 +120,7 @@ function App() {
       <div id="minicircle"></div>
       <div id="main" data-scroll-container data-scroll-speed="2">
       <Navbar showMenu={showMenu} setShowMenu={setShowMenu} mobile={mobileMenu} toggleMenu={setMobileMenu} id="nav"></Navbar>
-      <div id="hero">
-            <HeroBanner></HeroBanner>
-            <div id="chhotiheadings">
-                <div class="bounding">
-                    <h5 class="boundingelem moi">available for freelance</h5>
-                </div>
-                <div class="bounding">
-                    <h5 class="boundingelem moi oo">work for web</h5>
-                </div>
-            </div>
-            <Clients></Clients>
-            <div id="herofooter">
-                <a href="#">Working with MERN<FiArrowUpRight></FiArrowUpRight></a>
-                <a href="#">Developing Projects<FiArrowUpRight></FiArrowUpRight></a>
-                <div id="iconset">
-                    <div class="circle"> <IoArrowDownSharp></IoArrowDownSharp> </div>
-                    <div class="circle"> <IoArrowDownSharp></IoArrowDownSharp> </div>
-                </div>
-            </div>
-      </div>
-      <div id="second">
-        <div class="elem">
-          <h1>SKILLS</h1>
-          <h2>+</h2>
-          <div class="center14"></div>
-        </div>
-        <div class="elem">
-          <h1>SKILLS</h1>
-          <h2>+</h2>
-          <div class="center14"></div>
-        </div>
-        <div class="elem elemlast">
-          <h1>SKILLS</h1>
-          <h2>+</h2>
-          <div class="center14"></div>
-        </div>
-      </div>
-      <About src={loh}></About>
+      <Hero></Hero>
       <Footer dateState={dateState}></Footer>
     </div>
     {mobileMenu && (
