+++ src/transition/style.css
@@ -57,7 +57,7 @@
   
     }
     100% {
-      color: rgba(159, 242, 35, 0.897);
+      /* color: rgba(159, 242, 35, 0.897); */
         height: 100%;
       
     }
